moved chart generation to new file
